requirements.txt
streamlit==1.31.0
streamlit-lottie
pymongo==4.6.1
google-generativeai==0.3.2
bcrypt==4.1.2
plotly==5.18.0
fpdf==1.7.2
python-dotenv